mv GatewayAppBinding to gateway module
