+++ src/dashboard/apigateway/apigateway/apps/gateway/admin.py
@@ -15,32 +15,15 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
-# Generated by Django 3.2.18 on 2023-09-01 07:25
+from django.contrib import admin
 
-from django.db import migrations, models
-import django.db.models.deletion
+from .models import GatewayAppBinding
 
 
-class Migration(migrations.Migration):
+class GatewayAppBindingAdmin(admin.ModelAdmin):
+    list_display = ["id", "gateway", "bk_app_code", "updated_time"]
+    search_fields = ["gateway__id", "bk_app_code"]
+    list_filter = ["gateway"]
 
-    dependencies = [
-        ('core', '0036_proxy_backend'),
-        ('label', '0003_auto_20230227_2006'),
-    ]
 
-    operations = [
-        migrations.RenameField(
-            model_name='apilabel',
-            old_name='api',
-            new_name='gateway',
-        ),
-        migrations.AlterField(
-            model_name='apilabel',
-            name='gateway',
-            field=models.ForeignKey(db_column='api_id', on_delete=django.db.models.deletion.CASCADE, to='core.gateway'),
-        ),
-        migrations.AlterUniqueTogether(
-            name='apilabel',
-            unique_together={('gateway', 'name')},
-        ),
-    ]
+admin.site.register(GatewayAppBinding, GatewayAppBindingAdmin)
