+++ src/dashboard/apigateway/apigateway/apps/gateway/apps.py
@@ -15,32 +15,8 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
-# Generated by Django 3.2.18 on 2023-09-01 07:25
-
-from django.db import migrations, models
-import django.db.models.deletion
+from django.apps import AppConfig
 
 
-class Migration(migrations.Migration):
-
-    dependencies = [
-        ('core', '0036_proxy_backend'),
-        ('label', '0003_auto_20230227_2006'),
-    ]
-
-    operations = [
-        migrations.RenameField(
-            model_name='apilabel',
-            old_name='api',
-            new_name='gateway',
-        ),
-        migrations.AlterField(
-            model_name='apilabel',
-            name='gateway',
-            field=models.ForeignKey(db_column='api_id', on_delete=django.db.models.deletion.CASCADE, to='core.gateway'),
-        ),
-        migrations.AlterUniqueTogether(
-            name='apilabel',
-            unique_together={('gateway', 'name')},
-        ),
-    ]
+class GatewayConfig(AppConfig):
+    name = "apigateway.apps.gateway"
