+++ src/dashboard/apigateway/apigateway/apps/gateway/models.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 #
 # TencentBlueKing is pleased to support the open source community by making
 # 蓝鲸智云 - API 网关(BlueKing - APIGateway) available.
@@ -15,22 +16,24 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
-# Generated by Django 3.2.18 on 2023-08-31 11:14
+from django.db import models
 
-from django.db import migrations, models
-import django.db.models.deletion
+from apigateway.common.mixins.models import OperatorModelMixin, TimestampedModelMixin
+from apigateway.core.models import Gateway
 
 
-class Migration(migrations.Migration):
+class GatewayAppBinding(TimestampedModelMixin, OperatorModelMixin):
+    """
+    网关绑定的蓝鲸应用
+    - 仅影响 HomePage 中运维开发分数的计算
+    """
 
-    dependencies = [
-        ('core', '0036_auto_20230824_1737'),
-    ]
+    gateway = models.ForeignKey(Gateway, on_delete=models.CASCADE)
+    bk_app_code = models.CharField(max_length=32, db_index=True)
 
-    operations = [
-        migrations.AddField(
-            model_name='proxy',
-            name='backend',
-            field=models.ForeignKey(default=None, null=True, on_delete=django.db.models.deletion.PROTECT, to='core.backend'),
-        ),
-    ]
+    def __str__(self):
+        return f"<GatewayAppBinding: {self.bk_app_code}/{self.gateway_id}>"
+
+    class Meta:
+        db_table = "core_gateway_app_binding"
+        unique_together = ("gateway", "bk_app_code")
