+++ src/dashboard/apigateway/apigateway/conf/default.py
@@ -88,6 +88,7 @@ INSTALLED_APPS = [
     "apigateway.apps.monitor",
     "apigateway.schema",
     "apigateway.core",
+    "apigateway.apps.gateway",
     "apigateway.apps.access_strategy",
     "apigateway.apps.plugin",
     "apigateway.apps.label",
