+++ src/dashboard/apigateway/apigateway/tests/biz/test_gateway_app_binding.py
@@ -16,8 +16,8 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
+from apigateway.apps.gateway.models import GatewayAppBinding
 from apigateway.biz.gateway_app_binding import GatewayAppBindingHandler
-from apigateway.core.models import GatewayAppBinding
 
 
 class TestGatewayAppBindingHandler:
